refactor: export rules and configs using `requireindex`
