+++ lib/rules/_syntaxes/dynamic-directive-arguments.js
