+++ lib/rules/_syntaxes/scope-attribute.js
