+++ lib/rules/_syntaxes/script-setup.js
