+++ lib/rules/_syntaxes/slot-attribute.js
