+++ lib/rules/_syntaxes/slot-scope-attribute.js
