+++ lib/rules/_syntaxes/style-css-vars-injection.js
