+++ lib/rules/_syntaxes/utils/can-convert-to-v-slot.js
