+++ lib/rules/_syntaxes/v-bind-attr-modifier.js
