+++ lib/rules/_syntaxes/v-bind-prop-modifier-shorthand.js
