+++ lib/rules/_syntaxes/v-is.js
