+++ lib/rules/no-deprecated-scope-attribute.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const utils = require('../utils')
-const scopeAttribute = require('./syntaxes/scope-attribute')
+const scopeAttribute = require('./_syntaxes/scope-attribute')
 
 module.exports = {
   meta: {
