+++ lib/rules/no-deprecated-slot-attribute.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const utils = require('../utils')
-const slotAttribute = require('./syntaxes/slot-attribute')
+const slotAttribute = require('./_syntaxes/slot-attribute')
 
 module.exports = {
   meta: {
