+++ lib/rules/no-deprecated-slot-scope-attribute.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const utils = require('../utils')
-const slotScopeAttribute = require('./syntaxes/slot-scope-attribute')
+const slotScopeAttribute = require('./_syntaxes/slot-scope-attribute')
 
 module.exports = {
   meta: {
