+++ lib/rules/no-deprecated-v-is.js
@@ -5,7 +5,7 @@
 'use strict'
 
 const utils = require('../utils')
-const vIs = require('./syntaxes/v-is')
+const vIs = require('./_syntaxes/v-is')
 
 module.exports = {
   meta: {
