+++ lib/rules/no-unsupported-features.js
@@ -16,22 +16,22 @@ const utils = require('../utils')
 
 const FEATURES = {
   // Vue.js 2.5.0+
-  'slot-scope-attribute': require('./syntaxes/slot-scope-attribute'),
+  'slot-scope-attribute': require('./_syntaxes/slot-scope-attribute'),
   // Vue.js 2.6.0+
-  'dynamic-directive-arguments': require('./syntaxes/dynamic-directive-arguments'),
-  'v-slot': require('./syntaxes/v-slot'),
+  'dynamic-directive-arguments': require('./_syntaxes/dynamic-directive-arguments'),
+  'v-slot': require('./_syntaxes/v-slot'),
   // Vue.js 3.0.0+
-  'v-model-argument': require('./syntaxes/v-model-argument'),
-  'v-model-custom-modifiers': require('./syntaxes/v-model-custom-modifiers'),
-  'v-is': require('./syntaxes/v-is'),
-  'script-setup': require('./syntaxes/script-setup'),
-  'style-css-vars-injection': require('./syntaxes/style-css-vars-injection'),
+  'v-model-argument': require('./_syntaxes/v-model-argument'),
+  'v-model-custom-modifiers': require('./_syntaxes/v-model-custom-modifiers'),
+  'v-is': require('./_syntaxes/v-is'),
+  'script-setup': require('./_syntaxes/script-setup'),
+  'style-css-vars-injection': require('./_syntaxes/style-css-vars-injection'),
   // Vue.js 3.1.0+
-  'is-attribute-with-vue-prefix': require('./syntaxes/is-attribute-with-vue-prefix'),
+  'is-attribute-with-vue-prefix': require('./_syntaxes/is-attribute-with-vue-prefix'),
   // Vue.js 3.2.0+
-  'v-memo': require('./syntaxes/v-memo'),
-  'v-bind-prop-modifier-shorthand': require('./syntaxes/v-bind-prop-modifier-shorthand'),
-  'v-bind-attr-modifier': require('./syntaxes/v-bind-attr-modifier')
+  'v-memo': require('./_syntaxes/v-memo'),
+  'v-bind-prop-modifier-shorthand': require('./_syntaxes/v-bind-prop-modifier-shorthand'),
+  'v-bind-attr-modifier': require('./_syntaxes/v-bind-attr-modifier')
 }
 
 const SYNTAX_NAMES = /** @type {(keyof FEATURES)[]} */ (Object.keys(FEATURES))
