+++ package.json
@@ -55,6 +55,7 @@
   "dependencies": {
     "eslint-utils": "^2.1.0",
     "natural-compare": "^1.4.0",
+    "requireindex": "^1.2.0",
     "semver": "^6.3.0",
     "vue-eslint-parser": "^7.10.0"
   },
